# server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/auth/google":{"post":{"operationId":"AuthController_googleAuth","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleAuthDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/otp":{"post":{"operationId":"AuthController_otpAuth","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpAuthDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/otp/verify":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/chats/{id}/messages":{"post":{"operationId":"ChatController_createMessage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Chat"]}},"/api/chats":{"get":{"operationId":"ChatController_getChats","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Chat"]}},"/api/chats/new":{"post":{"operationId":"ChatController_createChat","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Chat"]}},"/api/chats/{id}":{"get":{"operationId":"ChatController_getChat","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Chat"]},"patch":{"operationId":"ChatController_updateChat","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateChatDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Chat"]},"delete":{"operationId":"ChatController_deleteChat","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Chat"]}},"/api/chats/upload":{"post":{"operationId":"ChatController_uploadFiles","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Chat"]}}},"info":{"title":"NestJS API","description":"API for NestJS","version":"1.0","contact":{}},"tags":[{"name":"API","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"GoogleAuthDto":{"type":"object","properties":{"idToken":{"type":"string","description":"The ID token from Google"}},"required":["idToken"]},"OtpAuthDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"VerifyOtpDto":{"type":"object","properties":{"email":{"type":"string"},"otp":{"type":"string"}},"required":["email","otp"]},"MessageDto":{"type":"object","properties":{"message":{"type":"object","description":"The message content and metadata"},"model":{"type":"string","description":"Model to use in format: provider/model-name (e.g., openai/gpt-4o-mini, anthropic/claude-3-5-sonnet-20241022, google/gemini-1.5-pro)","example":"openai/gpt-4o-mini"}},"required":["message","model"]},"UpdateChatDto":{"type":"object","properties":{}}}}}